Register a before_request event for all blueprints, that defines the authentication requirement. There are three authentication methods:  - requires_no_auth - public endpoint that does not require an Authorisation header  - requires_auth - public endpoints that need an API key in the Authorisation header  - requires_admin_auth - private endpoint that requires an Authorisation header which contains the API key for the defined as the client admin user
