+++ app/notifications/notifications_sms_callback.py
@@ -1,0 +1,47 @@
+from flask import Blueprint
+from flask import json
+from flask import request, jsonify
+
+from app.errors import InvalidRequest, register_errors
+from app.notifications.process_client_response import validate_callback_data, process_sms_client_response
+
+sms_callback_blueprint = Blueprint("sms_callback_blueprint", __name__, url_prefix="/notifications/sms")
+register_errors(sms_callback_blueprint)
+
+
+@sms_callback_blueprint.route('/mmg', methods=['POST'])
+def process_mmg_response():
+    client_name = 'MMG'
+    data = json.loads(request.data)
+    errors = validate_callback_data(data=data,
+                                    fields=['status', 'CID'],
+                                    client_name=client_name)
+    if errors:
+        raise InvalidRequest(errors, status_code=400)
+
+    success, errors = process_sms_client_response(status=str(data.get('status')),
+                                                  reference=data.get('CID'),
+                                                  client_name=client_name)
+    if errors:
+        raise InvalidRequest(errors, status_code=400)
+    else:
+        return jsonify(result='success', message=success), 200
+
+
+@sms_callback_blueprint.route('/firetext', methods=['POST'])
+def process_firetext_response():
+    client_name = 'Firetext'
+    errors = validate_callback_data(data=request.form,
+                                    fields=['status', 'reference'],
+                                    client_name=client_name)
+    if errors:
+        raise InvalidRequest(errors, status_code=400)
+
+    status = request.form.get('status')
+    success, errors = process_sms_client_response(status=status,
+                                                  reference=request.form.get('reference'),
+                                                  client_name=client_name)
+    if errors:
+        raise InvalidRequest(errors, status_code=400)
+    else:
+        return jsonify(result='success', message=success), 200
